refactor: extract teamswrapper component
