+++ src/app/shell/teams/page.tsx
@@ -1,46 +1,6 @@
 'use client'
 
-import { Button, PageHeading, Team } from '@/components'
-import { UserProfile } from '@/types'
-
-const userProfiles: UserProfile[] = [
-  {
-    avatar_id: null,
-    bio: 'Team Leader',
-    created_at: '2021-08-18T18:00:00.000Z',
-    first_name: 'John',
-    id: '1',
-    last_name: 'Doe',
-    user_id: '1234',
-  },
-  {
-    avatar_id: null,
-    bio: 'Team Leader',
-    created_at: '2021-08-18T18:00:00.000Z',
-    first_name: 'John',
-    id: '1',
-    last_name: 'Doe',
-    user_id: '1234',
-  },
-  {
-    avatar_id: null,
-    bio: 'Team Leader',
-    created_at: '2021-08-18T18:00:00.000Z',
-    first_name: 'John',
-    id: '1',
-    last_name: 'Doe',
-    user_id: '1234',
-  },
-  {
-    avatar_id: null,
-    bio: 'Team Leader',
-    created_at: '2021-08-18T18:00:00.000Z',
-    first_name: 'John',
-    id: '1',
-    last_name: 'Doe',
-    user_id: '1234',
-  },
-]
+import { Button, PageHeading, TeamsWrapper } from '@/components'
 
 export default function Teams() {
   return (
@@ -50,8 +10,7 @@ export default function Teams() {
         subtitle="Manage your teams, members, and team roles."
       />
       <Button onClick={() => {}}>Create new team</Button>
-      <Team members={userProfiles} name="California Team" />
-      <Team members={[]} name="Los Angeles Team" />
+      <TeamsWrapper />
     </div>
   )
 }
